requirements.txt
aiohttp==3.8.3
aiosignal==1.2.0
APScheduler==3.6.3
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
backports.zoneinfo==0.2.1
base58==2.1.1
bitarray==2.6.0
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
cytoolz==0.12.0
eth-abi==2.2.0
eth-account==0.5.9
eth-hash==0.5.0
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.9.5
frozenlist==1.3.1
hexbytes==0.3.0
idna==3.4
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipfshttpclient==0.8.0a2
jsonschema==4.16.0
lru-dict==1.1.8
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
numpy==1.21.6
pandas==1.3.5
parsimonious==0.8.1
pkgutil_resolve_name==1.3.10
protobuf==3.20.2
psycopg2-binary==2.9.3
pycryptodome==3.15.0
pyrsistent==0.18.1
python-dateutil==2.8.2
python-telegram-bot==13.14
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
rlp==2.0.1
six==1.16.0
toolz==0.12.0
tornado==6.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
varint==1.0.2
web3==5.31.0
websockets==9.1
yarl==1.8.1
zipp==3.8.1
